hypertension/scripts/initial_deployment/pyproject.toml
[tool.poetry]
name = "deploy-wkhtmltopdf"
version = "0.0.1"
description = "Deploy wkhtmltopdf AmazonLinux2 binary layer"
authors = ["Julia Allen <julia.l.allen@gsa.gov"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.17.87"
requests = "^2.26.0"
"ruamel.yaml" = "^0.17.10"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-watch = "^4.2.0"
python-dotenv = "^0.17.1"
pytest-coverage = "*"
flake8 = "^3.9.2"
pylint = "^2.8.2"
black = "^21.5b2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.pylint."MESSAGES CONTROL"]
enable = [
    "use-symbolic-message-instead",
    "useless-supression",
    "fixme",
]
disable = [
    "attribute-defined-outside-init",
    "duplicate-code",
    "invalid-name",
    "missing-docstring",
    "protected-access",
    "too-few-public-methods",
    "import-error",
    "no-value-for-parameter",
    "format",
    "pointless-string-statement",
    "unnecessary-lambda",
]
